+++ BugsnagPerformanceSwiftUI/BugsnagPerformanceSwiftUI.h
@@ -1,0 +1,19 @@
+//
+//  BugsnagPerformanceSwiftUI.h
+//  BugsnagPerformanceSwiftUI
+//
+//  Created by Karl Stenerud on 05.03.24.
+//  Copyright © 2024 Bugsnag. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for BugsnagPerformanceSwiftUI.
+FOUNDATION_EXPORT double BugsnagPerformanceSwiftUIVersionNumber;
+
+//! Project version string for BugsnagPerformanceSwiftUI.
+FOUNDATION_EXPORT const unsigned char BugsnagPerformanceSwiftUIVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <BugsnagPerformanceSwiftUI/PublicHeader.h>
+
+