+++ Sources/BugsnagPerformanceSwiftUI/BugsnagPerformanceSwiftUIInstrumentation.swift
@@ -1,0 +1,1 @@
+../BugsnagPerformanceSwift/BugsnagPerformanceSwiftUIInstrumentation.swift